FEAT : 관리자 기능 개발 중

+++ src/main/java/com/meta/overwash/mapper/ReservationConfirmedMapper.java
@@ -5,14 +5,14 @@ import java.util.List;
 import com.meta.overwash.domain.ReservationConfirmedDTO;
 
 public interface ReservationConfirmedMapper {
-	
+
 	public List<ReservationConfirmedDTO> getList();
 
 	public void insertReservationConfirm(ReservationConfirmedDTO rc);
-	
-	public ReservationConfirmedDTO getReservationConfirm(Long confirmId); 
+
+	public ReservationConfirmedDTO getReservationConfirm(Integer confirmId);
 
 	public int updateReservationConfirm(ReservationConfirmedDTO rc);
-	
-	public int deleteReservationConfirm(Long confirmId); 
+
+	public int deleteReservationConfirm(Long confirmId);
 }
